pyproject.toml
[project]
name = "reddit-memes"
version = "0.0.1"
description = "This project aims to predict meme success on Reddit with a multimodal approach using computer vision, generative AI and other techniques."
authors = [{ name = "Francisco Gomes", email = "lnpg1999@gmail.com" }, { name = "Guilherme Raimundo", email = "58053@novasbe.pt" }, { name = "Marta Dinis", email = "43027@novasbe.pt" }, { name = "Martim Costa", email = "martimbdcosta@gmail.com" }, { name = "Martim Esteves", email = "46953@novasbe.pt" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.12,<3.13"
dependencies = [
    "dask>=2024.9.0",
    "dask-ml>=2024.4.4",
    "ipykernel>=6.29.5",
    "lightgbm>=4.5.0",
    "matplotlib>=3.9.2",
    "networkx>=3.3",
    "nltk>=3.9.1",
    "pandas>=2.2.2",
    "pillow>=10.4.0",
    "pipdeptree>=2.23.3",
    "pyarrow>=17.0.0",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "transformers>=4.44.2",
    "xgboost>=2.1.1",
    "zstandard>=0.23.0",
    "types-requests>=2.32.0.20240914",
    "types-pillow>=10.2.0.20240822",
    "types-tensorflow>=2.17.0.20240820",
    "pandas-stubs>=2.2.2.240909",
    "urllib3>=2.2.3",
    "deepface>=0.0.93",
    "tf-keras>=2.16.0",
    "tensorflow[and-cuda]==2.16.1",
    "ollama>=0.3.3",
    "tensorflow-hub>=0.16.1",
    "statsmodels>=0.14.4",
    "stargazer>=0.0.7",
    "python-docx>=1.1.2",
    "beautifulsoup4>=4.12.3",
]

[project.urls]
Homepage = "https://martimesteves1.github.io/reddit-memes/"
Repository = "https://github.com/martimesteves1/reddit-memes"
Documentation = "https://martimesteves1.github.io/reddit-memes/"

[tool.uv]
dev-dependencies = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "deptry>=0.20.0",
    "mypy>=0.991",

    "ruff>=0.0.235",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.19.0",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["site"]

[tool.mypy]
files = ["reddit_memes"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
plugins = ["numpy.typing.mypy_plugin"]

[[tool.mypy.overrides]]
module = ["dask.*", "distributed.*", "dask_ml.*", "sklearn.*", "xgboost.*", "lightgbm.*", "transformers.*", "nltk.*", "networkx.*", "scipy.*", "zstandard.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 79
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
lint.ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true



[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
